V2.0.14
Reworked the way pages work.

you now create pages as you need them and can give/rename them.

also added a list box to quick select the named page by clicking on it.

future plans are to make it so you can add/remove hotbars to the page instead of just having the default 10 you will be able to have as few or many as you need on that page.