V2.0.205
added throttle setting for hotbar/page changes

this allows users to adjust it if they are having issues with the hotbars getting messed up because i/o isnt keeping up with scrolling.

the other way for me to handle this involves threading and i want to avoid that as it's much more complicated and too much work for almost no benefit.